Fix build warning for the below build option
export CFLAGS+="-Wall -Wcast-align -Wcast-qual -Wextra -Wno-array-bounds -Wno-empty-body -Wno-ignored-qualifiers -Wno-unused-parameter -Wshadow -Wwrite-strings -Wswitch-default -fno-common -fno-omit-frame-pointer -fno-optimize-sibling-calls -fno-strict-aliasing -fno-unroll-loops -fsigned-char -fstrict-overflow -Wno-unused-but-set-parameter -Wno-unused-but-set-variable"

export CXXFLAGS+="-Wall -Wcast-align -Wcast-qual -Wextra -Wnon-virtual-dtor -Wno-array-bounds -Wno-c++0x-compat -Wno-empty-body -Wno-ignored-qualifiers -Wno-unused-parameter -Wshadow -Wwrite-strings -Wswitch-default -fno-omit-frame-pointer -fno-optimize-sibling-calls -fno-strict-aliasing -fno-unroll-loops -fsigned-char -fstrict-overflow -Wno-unused-but-set-parameter -Wno-conversion-null -Wno-unused-but-set-variable"

+++ ism/extras/efl_panel/iseselector.cpp
@@ -116,7 +116,7 @@ gl_ise_selected_cb (void *data, Evas_Object *obj, void *event_info)
 
     LOGD ("selected item : %d\n", index);
 
-    if (index >= (int)_uuids.size ())
+    if (index >= _uuids.size ())
         return;
 
     if (_ise_selector_radio_grp)
