+++ ism/utils/scim_x11_utils.cpp
@@ -195,7 +195,7 @@ scim_x11_keymask_x11_to_scim (Display *display, unsigned int xkeystate)
     initialize_modifier_bits (display);
 
     // Check Meta mask first, because it's maybe a mask combination.
-    if (__current_meta_mask && (xkeystate & __current_meta_mask) == __current_meta_mask) {
+    if (__current_meta_mask && (int)(xkeystate & __current_meta_mask) == __current_meta_mask) {
         mask |= SCIM_KEY_MetaMask;
         xkeystate &= ~__current_meta_mask;
     }
@@ -215,22 +215,22 @@ scim_x11_keymask_x11_to_scim (Display *display, unsigned int xkeystate)
         xkeystate &= ~ControlMask;
     }
 
-    if (__current_alt_mask && (xkeystate & __current_alt_mask) == __current_alt_mask) {
+    if (__current_alt_mask && ((int)xkeystate & __current_alt_mask) == __current_alt_mask) {
         mask |= SCIM_KEY_AltMask;
         xkeystate &= ~__current_alt_mask;
     }
 
-    if (__current_super_mask && (xkeystate & __current_super_mask) == __current_super_mask) {
+    if (__current_super_mask && ((int)xkeystate & __current_super_mask) == __current_super_mask) {
         mask |= SCIM_KEY_SuperMask;
         xkeystate &= ~__current_super_mask;
     }
 
-    if (__current_hyper_mask && (xkeystate & __current_hyper_mask) == __current_hyper_mask) {
+    if (__current_hyper_mask && ((int)xkeystate & __current_hyper_mask) == __current_hyper_mask) {
         mask |= SCIM_KEY_HyperMask;
         xkeystate &= ~__current_hyper_mask;
     }
 
-    if (__current_numlock_mask && (xkeystate & __current_numlock_mask) == __current_numlock_mask) {
+    if (__current_numlock_mask && (int)(xkeystate & __current_numlock_mask) == __current_numlock_mask) {
         mask |= SCIM_KEY_NumLockMask;
     }
